Finishes dog walk api

+++ pages/api/dog-walks/[dogWalkId].ts
@@ -7,15 +7,33 @@ export default async function handler(
   res: NextApiResponse
 ) {
   switch (req.method) {
+    case 'PUT':
+      return handlePut(req, res);
     case 'DELETE':
-      return deleteHandler(req, res);
+      return handleDelete(req, res);
     default:
-      res.status(405);
+      res.status(405).end();
   }
 }
 
-async function deleteHandler(req: NextApiRequest, res: NextApiResponse) {
+async function handlePut(req: NextApiRequest, res: NextApiResponse) {
   const { dogWalkId } = req.query;
+  if (typeof dogWalkId !== 'string') {
+    return res.status(400).end();
+  }
+  if (dogWalkId !== req.body.id) {
+    return res.status(400).end();
+  }
+  const updatedDogWalk = req.body;
+  const result = await dogWalkService.updateDogWalk(updatedDogWalk);
+  res.status(200).json(result);
+}
+
+async function handleDelete(req: NextApiRequest, res: NextApiResponse) {
+  const { dogWalkId } = req.query;
+  if (typeof dogWalkId !== 'string') {
+    return res.status(400).end();
+  }
   const result = await dogWalkService.deleteDogWalk(dogWalkId);
   res.status(200).json({ id: dogWalkId });
 }
